+++ mne/simulation/__init__.py
@@ -3,5 +3,6 @@
 
 from .evoked import add_noise_evoked, simulate_evoked, simulate_noise_evoked
 from .raw import simulate_raw
-from .source import select_source_in_label, simulate_stc, simulate_sparse_stc
+from .source import (select_source_in_label, simulate_stc,
+                     simulate_sparse_stc)
 from .metrics import source_estimate_quantification